datastore is now an interface
One file at a time...
