+++ src/me/ryanhamshire/GriefPrevention/BlockEventHandler.java
@@ -23,6 +23,7 @@ import java.util.Collection;
 import java.util.List;
 
 import me.ryanhamshire.GriefPrevention.claim.Claim;
+import me.ryanhamshire.GriefPrevention.message.Messages;
 import me.ryanhamshire.GriefPrevention.player.PlayerData;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
