+++ src/me/ryanhamshire/GriefPrevention/PlayerRescueTask.java
@@ -18,6 +18,7 @@
  
 package me.ryanhamshire.GriefPrevention;
 
+import me.ryanhamshire.GriefPrevention.message.Messages;
 import me.ryanhamshire.GriefPrevention.player.PlayerData;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
