+++ src/me/ryanhamshire/GriefPrevention/message/Messages.java
@@ -16,7 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.ryanhamshire.GriefPrevention;
+package me.ryanhamshire.GriefPrevention.message;
 
 public enum Messages
 {
